exit test-polyfills workflow early if running from a forked repo.
This is because the workflow will fail due to not having access to the secrets required for browserstack to run.

We want to exit early because then we will not have this workflow run be marked as a failure. Having the workflow get marked as a failure will make a GitHub notifcation/email be sent to whoever follows the builds for polyfill-library (such as myself).