vendor mocha (#21)
`mocha` has a bunch of security issues but they are all in the node
path, not the browser path. It doesn't make sense for us to load this
from npm and to also pull in their ancient and vulnerable dependencies.

I've created a version where only the runners, reporters, ... exist that
we need.
I've also remove the usual B.S. that people seem to like, colored
outputs, ...

I might go further later and create a single file manually without all
the extension points that `mocha` offers. We don't need those extension
points if we will never use them.

That should produce something that can be more easily maintained from
here.
But this version will do for now.